pyproject.toml
[project]
name = "mlgaze-viewer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pandas-stubs==2.3.0.250703",
    "pyyaml>=6.0.2",
    "rerun-sdk>=0.24.0",
    "textual>=5.2.0",
]
